Logic changes
I noticed AHK bugging out, buttons were getting stuck down. Now I can't
reproduce it.  Not sure if the changes introduced in this commit caused
AHK to be sad, or if AHK is always sad and I only got a taste of it for
that brief window. Friends don't let friends program in AHK...